Qt-ify KaraokeContainer and KaraokeData
Gets rid of the need to convert between std::string and QString,
simplifies handing text encodings, and lets us use signals and
slots in Song in the future.

The original intention was that KaraokeContainer and KaraokeData
wouldn't use Qt so that they would be easier to use as libraries
in other programs, but that makes things more impractical for us,
Qt was already required for XML parsing anyway, and I'm not sure
if there's anyone who actually wants to use this code as a library.