Use QAudioOutput with dr_mp3 and dr_flac instead of QMediaPlayer (which sucks) (#3)
* Audio playback with QMediaPlayer (limited compatibility)

* Proof of concept of using dr_mp3 to decode MP3

(QMediaPlayer really sucks at playing normal MP3s consistently, so
the idea here is to use an external library to decode for it.
Unfortunately the decoded WAVE version has to be written to a
temporary file, because, well, QMediaPlayer really sucks.)

* Keep WAVE file in-memory instead of using tempfile

It turns out that QIODevice does work with QMediaPlayer for a WAVE file,
but the fake path provided must end in .wav. I guess I had only tried
.mp3 fake paths. MIME type seems to be ignored. :(

* De-namespace AudioFile

* FLAC support

* Small fix, small cleanup

* Replace broken QMediaPlayer with QAudioOutput

It turns out QIODevice does not, in fact, work on macOS with QMediaPlayer like I thought it did. It was just my luck that the supposedly fake path I had chosen is where I had previously saved the .wav file to for testing. Oops.

Instead let's use QAudioOutput, which does at least work properly on macOS AFAICT, and means we don't have to pointlessly convert to WAV!

* Drop unnecessary spaces

* Use good™ license terms actually