SyllableDecorations: Calculate geometry on move, not paint
Should be faster, and seems to fix a bug where syllable decorations
wouldn't be visible when pasting many lines into the raw text view
while the rich text view was scrolled to a position that didn't
include the end of the document.

This necessitated making a change to LyricsEditor::SetMode to solve
a different problem that was masked before.